requirements.txt
airtable-python-wrapper
progress
wget